organize motion photo rust code
