+++ mobile/apps/photos/rust/src/api/usearch_api.rs
@@ -81,7 +81,6 @@ fn validate_semantic_search_exact_queries(
     Ok(())
 }
 
-
 fn flatten_semantic_search_exact_image_embeddings(
     image_embeddings: &[Vec<f32>],
     expected_dimension: Option<usize>,
@@ -201,7 +200,6 @@ pub struct SemanticSearchExactResponse {
     pub matches_per_query: Vec<Vec<SemanticSearchExactMatch>>,
 }
 
-
 #[frb(opaque)]
 pub struct SemanticSearchExactCache {
     image_file_ids: Vec<i64>,
@@ -258,7 +256,6 @@ pub fn semantic_search_exact(
         &minimum_similarities,
     )
 }
-
 
 #[frb(opaque)]
 pub struct VectorDB {
@@ -486,5 +483,4 @@ mod tests {
 
         assert!(err.contains("embedding dimension mismatch"));
     }
-
 }
